requirements.txt
Flask==3.0.0
Flask-SocketIO==5.3.6
Flask-CORS==4.0.0
python-socketio==5.9.0
python-engineio==4.7.1
eventlet==0.33.3
gunicorn==21.2.0
python-dotenv==1.0.0 